toggle comp
